requirements.txt
# Face Verification System - Production Requirements
# Compatible with Mac M2 (Apple Silicon) and CUDA environments

# Core ML/DL frameworks
torch>=1.13.0
torchvision>=0.14.0
facenet-pytorch>=2.5.3

# Image processing and computer vision
numpy>=1.21.0
Pillow>=9.0.0
opencv-python-headless>=4.5.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Machine learning and evaluation
scikit-learn>=1.0.0
scipy>=1.7.0

# Data handling and utilities
pandas>=1.3.0
tqdm>=4.64.0
pathlib>=1.0.0

# Development and testing
jupyter>=1.0.0
ipywidgets>=7.6.0
argparse

# Optional: For enhanced visualization
plotly>=5.0.0

# Note: torch will automatically detect and use:
# - MPS (Metal Performance Shaders) on Apple Silicon
# - CUDA on NVIDIA GPUs  
# - CPU fallback on other systems
